// src/contracts/types.ts
export type SupportedNetworkId =
  | "1"
  | "5"
  | "11155111"
  | "1337"
  | "168587773"
  | "81457";

export type ContractName =
  | "BaseRegistrarImplementation"
  | "ETHRegistrarController"
  | "Multicall"
  | "NameWrapper"
  | "DNSRegistrar"
  | "PublicResolver"
  | "ENSRegistry"
  | "ReverseRegistrar"
  | "UniversalResolver"
  | "BulkRenewal"
  | "Voucher"
  | "Referral";

export type ContractNameAlgo =
  | "ERC20"
  | "MergedPool"
  | "IUniswapV2Pair"
  | "Bank"
  | "Vesting"
  | "Treasury";